Reworked app state distribution mechanism
Now the "app-component" is the only component allowed to deref the app
state; the relevant pieces are passed down to the sub-components. Now
the sub components are only being re-rendered if and only if their
relevant pieces of the state are changed. Previously the entire app was
being re-rendered anytime the state was changed at all.

https://youtu.be/oRmj3IUkRVk?t=28m

The items and champs, however, are now in their own atoms, since it
seems okay to me to cause global re-renders twice at the beginning, as
the data is loaded.